requirements.txt
# --------- pytorch --------- #
torch
torchvision
pytorch-lightning>=1.5.0
lightning-bolts
torchmetrics
torchattacks
#segmentation-models-pytorch
albumentations
einops
setuptools==68.2.2 # Fix version install
pathlib2

# --------- hydra --------- #
hydra-core==1.1.0
hydra-colorlog
# hydra-optuna-sweeper==1.1.0
# hydra-ax-sweeper
# hydra-ray-launcher
# hydra-submitit-launcher

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml
# torch_tb_profiler
clearml

# --------- linters --------- #
#pre-commit      # hooks for applying linters on commit
#black           # code formatting
#isort           # import sorting
#flake8          # code analysis

# --------- others --------- #
python-dotenv   # loading env variables from .env file
rich            # beautiful text formatting in terminal
pytest          # tests
sh              # for running bash commands in some tests
scikit-learn    # used in some callbacks
seaborn         # used in some callbacks
jupyterlab      # better jupyter notebooks
pudb            # debugger
